Lecturer Report (outputs/console)
The class LecturerReport in the CMSApp.The CMSApp package gives the Course Management System (CMS) a unique feature that lets you focus on the work and efforts of teachers. This part is very important for showing how influential and involved the academic staff is in the school because it lets you make detailed reports about lecturers that include their names, titles, areas of expertise, modules taught, and the number of students they've reached. The exporting ways (outputToTxtFile, outputToCsvFile, and outputToConsole) include all the features needed to change the query results into forms that are easy to use. These methods show how to handle files and show data in Java in a useful way. Any software worker who works with data-heavy apps needs to know these skills.