Admin
This is the Admin class in the CMS app.The CMSApp package is an important part of the Course Management System (CMS). It's made to hold the features and functions that management users need. As a subclass of the User entity, it gets basic properties like login and password, but it stands out by having extra features that are important for system management, like the ability to manage user accounts and personal credentials.

In the CMS, this class keeps a basic list of User objects that reflect all of the users. This choice in design lets the Admin class do important things directly on the list of users, like adding new users, changing information about current users, and removing users from the system. When it comes to keeping the CMS's user group honest and up to date, these features are essential.

To be more specific, the addUser, modifyUser, deleteUser, and changeOwnCredentials class functions. Additionally, the setUsername and setPassword methods emphasise the class's focus on data consistency and encapsulation. They make sure that user traits are changed through controlled interfaces instead of direct access to fields, which is in line with object-oriented programming principles.

The Admin class is a big step towards making it possible to control users effectively in the CMS. It was added as part of project development. It sets the stage for advanced access control and administrative oversight, which are necessary parts of any complete management system.