Office
The CMS app's Office class.The CMSApp package includes features on the Course Management System (CMS) that are designed to help office workers. This shows how important the CMS is for organising and reporting on different academic aspects. This class, which is based on the base User class and is linked to the UserRole.OFFICE, is meant to make detailed records about classes, students, and teachers. These reports are an important part of managerial oversight because they help keep track of student records, course tasks, and teacher jobs. The methods generateCourseReport, generateStudentReport, and generateLecturerReport are just placeholders for now. When they are finally put into use, they will require a lot of complex data gathering and processing to give us useful information about the education environment. The Office class also shows how committed the system is to security and personalisation by giving office users ways to change their credentials, which protects the stability and privacy of access. By combining this class with InputUtils and the Main class, live report creation and user management will be possible. This will give office staff a dynamic and responsive user experience.