Lecturer
The Lecturer class in the cmsapp.CMSApp package is a custom job in the Course Management System (CMS) that focuses on features that are only available to teachers. It takes basic characteristics like login and password from the User class because it is an extension of the User class, but the method generateLecturerReport is what makes it unique. This method, which was created so that professors could easily gather and go over educational materials and lessons, is a key link between data and useful insights. It's just a placeholder right now, but its real goal is to collect and show information about the classes taught, how involved students are in their own learning, and maybe even how teachers affect students' success. Also, the class has a method that should let office workers change their credentials.