InputUtils class (Admin)
The CMSApp has a class called InputUtils.With the CMSApp package, you can handle user inputs across the Course Management System (CMS) like a tool belt. This class has a scanner for reading user input and gives an organised way to collect and handle data needed for many administrative jobs, like adding new people to the system.

In the addUser method, for example, InputUtils makes it easier to get the details needed to add a new user. It asks the admin for a new username, password, and job and gets them back, which is a simple and easy-to-understand way to connect. Not only does this way show how the class can help with managing users, it also stresses how important it is to validate data. It makes sure that only correct usernames and passwords are accepted by integrating with a DataValidator. This improves the accuracy and security of the system.

The InputUtils class is likely to grow a lot. It can grow overtime to include more complicated validation routines that can handle strict password rules and username limits. Its usefulness can also be increased by adding more CMS features, like the ability to make reports, update user information, and handle course entries, all through a uniform and easy-to-use interface. Also, making InputUtils work better with the Main class and other parts of the CMS could make the whole experience better for users. By handling and validating inputs in one place, the system can give users more timely feedback and help. This makes the CMS not only safer but also easier to access and use.