Main (add user)
When an administrative user calls the handleAdmin method, they are welcome and given a list of choices for handling other system users. Users can add new ones, change the information about current ones, delete users, change their own login information, and log out. This menu-based method makes sure that management jobs can be done quickly and correctly.

The purpose of the addUser method is to make it easy to add new people to the system. It asks the manager to enter important details like the new user's username, password, and job (manager, OFFICE, LECTURER). Once the data has been checked, a new User object is made and added to both a public userList and an admin-only list of people.