Admin (modify User)
The modifyUser method makes it easy for an administrator to change information about a Course Management System user. It asks the admin to enter the username of the user whose account they want to change, along with the new password and job for that user. This lets the admin make full changes to a user's account. When the method gets the new information, it uses the modifyUser function on the admin object and passes in the new username, password, and job. This changes the system's record for the given person right away to reflect the changes. At the end of the way, it confirms that the user's change was made successfully. This lets the admin know exactly what was done. This feature is very important for making sure that user info in the CMS is correct and useful.