InputUtils (Modify User by admin)
The modifyUser method is meant to change information about a current user in the Course Management System. Password validation adds an extra layer of security. If the manager wants to change a user's job, they need to enter their new username, password, and role (manager, OFFICE, LECTURER). The DataValidator is then used to make sure that the new password is correct. The user's information is changed through the modifyUser method of the admin object if the password meets the approval conditions. When the change is made successfully, a message is shown. An error message is shown if the new password fails approval, letting you know that the password is not valid. This process makes sure that both the security standards are followed and that user info is kept safe.