Main (delete User/change own credentials)
By giving the username of the user, the deleteUser way lets an administrator get rid of that person from the Course Management System. Once the username is given, the deleteUser function on the admin object is called to remove the user. This is followed by a message confirming that the deletion was successful.

An admin can change their username and password using the changeOwnCredentials method. It asks the admin for a new username and password, then uses the changeOwnCredentials function on the admin object to make the changes. A message lets the admin know that the changes were made successfully. Both ways improve the system's security and managerial flexibility by making it easier to handle users and make changes to their credentials.