Input Utils (delete User/ change own credentials)
The deleteUser method asks an administrator for the username of the person they want to delete so that the user can be removed from the system. When the command is run, the given person is deleted, and a success message is shown.

The changeOwnCredentials method lets an admin change their login information after making sure the information is correct. It asks for a new account and password and uses a DataValidator to make sure the new password fits certain requirements. If the confirmation works, the admin's keys are changed and a message of success is shown. If not, a warning message will let the manager know that the password is wrong. This process shows how important security and data accuracy are to the system.