Main (Office functionalities)
As a dynamic interface for office staff in the Course Management System, the handleOffice method lets them do a variety of tasks that are specific to their management jobs. When users first log in, they are greeted and shown a list of things they can do, such as making course, student, and instructor reviews and updating their personal information. This way of doing things makes sure that office workers can quickly do their jobs, like getting to detailed academic records and keeping things safe by managing credentials. The logout choice, which ends the session and leaves the loop, shows that the method was designed with the user in mind, putting an emphasis on simplicity and ease of use.