Main (Lecturer)
In the Course Management System, the handleLecturer method gives teachers a special way to access and carry out tasks that are important to their job. When teachers log in, they are greeted and given a short list of choices. These include making reports about their teaching activities, updating their personal login information, and logging out of the system. This method is meant to keep things simple and quick for the user, so teachers can quickly choose an action without having to deal with extraneous details.

The process is built around a loop that keeps going until the speaker logs out. This lets them do more than one thing at once without having to restart the app. For input collection, the method uses the scanner object. Right after that, a line is used to handle possible input mismatches appropriately.